Samplers

RandomFeatures.Samplers.Sampler — Type

struct Sampler

Wraps the parameter distributions used to sample random features

parameter_distribution::EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution

A probability distribution, possibly with constraints

uniform_shift_bounds::Union{Nothing, AbstractVector}

Either nothing , or [lower bound, upper bound], which defines a uniform distribution used as a random shift

rng::Random.AbstractRNG

A random number generator state

RandomFeatures.Samplers.FeatureSampler — Function

FeatureSampler(
    parameter_distribution::EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution;
    uniform_shift_bounds,
    rng
) -> Sampler

basic constructor for a Sampler

RandomFeatures.Samplers.get_parameter_distribution — Function

get_parameter_distribution(
    s::Sampler
) -> EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution

gets the parameter_distribution field

RandomFeatures.Samplers.get_uniform_shift_bounds — Function

get_uniform_shift_bounds(
    s::Sampler
) -> Union{Nothing, AbstractVector}

gets the uniform_shift_bounds field

RandomFeatures.Samplers.get_rng — Function

get_rng(s::Sampler) -> Random.AbstractRNG

gets the rng field

StatsBase.sample — Method

sample(
    rng::Random.AbstractRNG,
    s::Sampler,
    n_draws::Int64
) -> EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution

samples the distribution within s, n_draws times using a random number generator rng. Can be called without rng (defaults to s.rng) or n_draws (defaults to 1)