Methods

RandomFeatures.Methods.RandomFeatureMethod — Type

struct RandomFeatureMethod{S<:AbstractString, USorM<:Union{AbstractMatrix, LinearAlgebra.UniformScaling}}

Holds configuration for the random feature fit

random_feature::RandomFeature: The random feature object

batch_sizes::Dict{S, Int64} where S<:AbstractString: A dictionary specifying the batch sizes. Must contain "train", "test", and "feature" keys

regularization::Union{AbstractMatrix, LinearAlgebra.UniformScaling}: A positive definite matrix used during the fit method to regularize the linear solve

tullio_threading::Bool: Use multithreading provided by Tullio

RandomFeatures.Methods.Fit — Type

struct Fit{V<:(AbstractVector), USorM<:Union{AbstractMatrix, LinearAlgebra.UniformScaling}}

Holds the coefficients and matrix decomposition that describe a set of fitted random features.

feature_factors::Decomposition: The LinearAlgreba matrix decomposition of (1 / m) * Feature^T * Feature + regularization

coeffs::AbstractVector: Coefficients of the fit to data

regularization::Union{AbstractMatrix, LinearAlgebra.UniformScaling}: feature-space regularization used during fit

RandomFeatures.Methods.get_random_feature — Function

get_random_feature(
    rfm::RandomFeatureMethod
) -> RandomFeature

gets the random_feature field

RandomFeatures.Methods.get_batch_sizes — Function

get_batch_sizes(
    rfm::RandomFeatureMethod
) -> Dict{S, Int64} where S<:AbstractString

gets the batch_sizes field

RandomFeatures.Methods.get_batch_size — Function

get_batch_size(
    rfm::RandomFeatureMethod,
    key::AbstractString
) -> Int64

get the specified batch size from batch_sizes field

RandomFeatures.Methods.get_regularization — Function

get_regularization(
    rfm::RandomFeatureMethod
) -> Union{AbstractMatrix, LinearAlgebra.UniformScaling}

gets the regularization field

get_regularization(
    f::Fit
) -> Union{AbstractMatrix, LinearAlgebra.UniformScaling}

gets the regularization field (note this is the feature-space regularization)

StatsBase.sample — Method

sample(rfm::RandomFeatureMethod) -> Any

samples the random_feature field

RandomFeatures.Methods.get_feature_factors — Function

get_feature_factors(f::Fit) -> Decomposition

gets the feature_factors field

RandomFeatures.Methods.get_coeffs — Function

get_coeffs(f::Fit) -> AbstractVector

gets the coeffs field

StatsAPI.fit — Function

fit(
    rfm::RandomFeatureMethod,
    input_output_pairs::EnsembleKalmanProcesses.DataContainers.PairedDataContainer;
    decomposition_type
) -> Fit

Fits a RandomFeatureMethod to input-output data, optionally provide a preferred LinearAlgebra matrix decomposition. Returns a Fit object.

StatsAPI.predict — Function

predict(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer;
    kwargs...
) -> Tuple{Any, Any}

Makes a prediction of mean and (co)variance of fitted features on new input data

StatsAPI.predict! — Function

predict!(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer,
    mean_store::AbstractMatrix{<:AbstractFloat},
    cov_store::AbstractArray{<:AbstractFloat, 3},
    buffer::AbstractArray{<:AbstractFloat, 3};
    kwargs...
)

Makes a prediction of mean and (co)variance of fitted features on new input data, overwriting the provided stores.

meanstore:`outputdimxn_samples`

covstore:`outputdimxoutputdimxnsamples`

buffer:n_samples x output_dim x n_features

RandomFeatures.Methods.predictive_mean — Function

predictive_mean(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer;
    kwargs...
) -> Tuple{Any, Any}

Makes a prediction of mean of fitted features on new input data. Returns a output_dim x n_samples array.

RandomFeatures.Methods.predictive_mean! — Function

predictive_mean!(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer,
    mean_store::Matrix{<:AbstractFloat};
    kwargs...
) -> Any

Makes a prediction of mean of fitted features on new input data. Writes into a provided output_dim x n_samples array: mean_store.

RandomFeatures.Methods.predictive_cov — Function

predictive_cov(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer,
    prebuilt_features::AbstractArray{<:AbstractFloat, 3};
    kwargs...
) -> Any

Makes a prediction of (co)variance of fitted features on new input data.

Returns a output_dim x output_dim x n_samples array

RandomFeatures.Methods.predictive_cov! — Function

predictive_cov!(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer,
    cov_store::AbstractArray{<:AbstractFloat, 3},
    buffer::AbstractArray{<:AbstractFloat, 3},
    prebuilt_features::AbstractArray{<:AbstractFloat, 3};
    tullio_threading,
    kwargs...
)

Makes a prediction of (co)variance of fitted features on new input data.

Writes into a provided output_dim x output_dim x n_samples array: cov_store, and uses provided n_samples x output_dim x n_features buffer.

RandomFeatures.Methods.predict_prior — Function

predict_prior(
    rfm::RandomFeatureMethod,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer;
    kwargs...
) -> Tuple{Any, Any}

Makes a prediction of mean and (co)variance with unfitted features on new input data

RandomFeatures.Methods.predict_prior_mean — Function

predict_prior_mean(
    rfm::RandomFeatureMethod,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer;
    kwargs...
) -> Tuple{Any, Any}

Makes a prediction of mean with unfitted features on new input data

RandomFeatures.Methods.predict_prior_cov — Function

predict_prior_cov(
    rfm::RandomFeatureMethod,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer;
    kwargs...
) -> Tuple{Any, Any}

Makes a prediction of (co)variance with unfitted features on new input data