Methods

RandomFeatures.Methods.RandomFeatureMethod — Type

struct RandomFeatureMethod

Holds configuration for the random feature fit

random_feature::RandomFeature

The random feature object

batch_sizes::Dict

A dictionary specifying the batch sizes. Must contain "train", "test", and "feature" keys

regularization::Real

A non-negative, additive regularization parameter used during the fit method in the linear solve

RandomFeatures.Methods.Fit — Type

struct Fit

Holds the coefficients and matrix decomposition that describe a set of fitted random features.

feature_factors::Decomposition

The LinearAlgreba matrix decomposition of (1 / m) * Feature^T * Feature + regularization * I

coeffs::AbstractVector

Coefficients of the fit to data

RandomFeatures.Methods.get_random_feature — Function

get_random_feature(
    rfm::RandomFeatureMethod
) -> RandomFeature

gets the random_feature field

RandomFeatures.Methods.get_batch_sizes — Function

get_batch_sizes(rfm::RandomFeatureMethod) -> Dict

gets the batch_sizes field

RandomFeatures.Methods.get_batch_size — Function

get_batch_size(
    rfm::RandomFeatureMethod,
    key::AbstractString
) -> Any

get the specified batch size from batch_sizes field

RandomFeatures.Methods.get_regularization — Function

get_regularization(rfm::RandomFeatureMethod) -> Real

gets the regularization field

StatsBase.sample — Method

sample(rfm::RandomFeatureMethod) -> Any

samples the random_feature field

RandomFeatures.Methods.get_feature_factors — Function

get_feature_factors(f::Fit) -> Decomposition

gets the feature_factors field

RandomFeatures.Methods.get_coeffs — Function

get_coeffs(f::Fit) -> AbstractVector

gets the coeffs field

StatsAPI.fit — Function

fit(
    rfm::RandomFeatureMethod,
    input_output_pairs::EnsembleKalmanProcesses.DataContainers.PairedDataContainer;
    decomposition_type
) -> Fit

Fits a RandomFeatureMethod to input-output data, optionally provide a preferred LinearAlgebra matrix decomposition. Returns a Fit object.

StatsAPI.predict — Function

predict(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer
) -> Tuple{Matrix{Float64}, Matrix{Float64}}

Makes a prediction of mean and (co)variance of fitted features on new input data

RandomFeatures.Methods.predictive_mean — Function

predictive_mean(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer
) -> Matrix{Float64}

Makes a prediction of mean of fitted features on new input data

RandomFeatures.Methods.predictive_cov — Function

predictive_cov(
    rfm::RandomFeatureMethod,
    fit::Fit,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer
) -> Tuple{Matrix{Float64}, Any}

Makes a prediction of (co)variance of fitted features on new input data

RandomFeatures.Methods.predict_prior — Function

predict_prior(
    rfm::RandomFeatureMethod,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer
) -> Tuple{Matrix{Float64}, Matrix{Float64}}

Makes a prediction of mean and (co)variance with unfitted features on new input data

RandomFeatures.Methods.predict_prior_mean — Function

predict_prior_mean(
    rfm::RandomFeatureMethod,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer
) -> Matrix{Float64}

Makes a prediction of mean with unfitted features on new input data

RandomFeatures.Methods.predict_prior_cov — Function

predict_prior_cov(
    rfm::RandomFeatureMethod,
    new_inputs::EnsembleKalmanProcesses.DataContainers.DataContainer
) -> Matrix{Float64}

Makes a prediction of (co)variance with unfitted features on new input data