Samplers

RandomFeatures.Samplers.Sampler — Type

struct Sampler{RNG<:Random.AbstractRNG}

Wraps the parameter distributions used to sample random features

parameter_distribution::EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution: A probability distribution, possibly with constraints

rng::Random.AbstractRNG: A random number generator state

RandomFeatures.Samplers.FeatureSampler — Function

FeatureSampler(
    parameter_distribution::EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution,
    bias_distribution::Union{Nothing, EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution};
    rng
) -> Sampler{Random._GLOBAL_RNG}

basic constructor for a Sampler

FeatureSampler(
    parameter_distribution::EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution,
    output_dim::Int64;
    uniform_shift_bounds,
    rng
) -> Sampler{Random._GLOBAL_RNG}

one can conveniently specify the bias as a uniform-shift uniform_shift_bounds with output_dim dimensions

RandomFeatures.Samplers.get_parameter_distribution — Function

get_parameter_distribution(
    s::Sampler
) -> EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution

gets the parameter_distribution field

RandomFeatures.Samplers.get_rng — Function

get_rng(s::Sampler) -> Random.AbstractRNG

gets the rng field

StatsBase.sample — Method

sample(
    rng::Random.AbstractRNG,
    s::Sampler,
    n_draws::Int64
) -> EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution

samples the distribution within s, n_draws times using a random number generator rng. Can be called without rng (defaults to s.rng) or n_draws (defaults to 1)