Utilities

Batching

RandomFeatures.Utilities.batch_generator — Function

batch_generator(
    array::AbstractArray,
    batch_size::Int64;
    dims
) -> Any

produces batched sub-array views of size batch_size along dimension dims.

Note

this creates views not copies. Modifying a batch will modify the original!

Matrix Decomposition

RandomFeatures.Utilities.Decomposition — Type

struct Decomposition{T, M<:(AbstractMatrix), MorF<:Union{AbstractMatrix, LinearAlgebra.Factorization}}

Stores a matrix along with a decomposition T=Factor, or pseudoinverse T=PseInv, and also computes the inverse of the Factored matrix (for several predictions this is actually the most computationally efficient action)

full_matrix::AbstractMatrix: The original matrix

decomposition::Union{AbstractMatrix, LinearAlgebra.Factorization}: The matrix decomposition, or pseudoinverse

inv_decomposition::AbstractMatrix: The matrix decomposition of the inverse, or pseudoinverse

RandomFeatures.Utilities.StoredInvType — Type

abstract type StoredInvType

Type used as a flag for the stored Decomposition type

RandomFeatures.Utilities.Factor — Type

abstract type Factor <: StoredInvType

RandomFeatures.Utilities.PseInv — Type

abstract type PseInv <: StoredInvType

RandomFeatures.Utilities.get_decomposition — Function

get_decomposition(
    d::Decomposition
) -> Union{AbstractMatrix, LinearAlgebra.Factorization}

get decomposition field

RandomFeatures.Utilities.get_full_matrix — Function

get_full_matrix(d::Decomposition) -> AbstractMatrix

get full_matrix field

RandomFeatures.Utilities.get_parametric_type — Function

get_parametric_type(
    d::Decomposition{T, M<:Union{AbstractMatrix, LinearAlgebra.Factorization}}
) -> Any

get the parametric type

RandomFeatures.Utilities.linear_solve — Function

linear_solve(
    d::Decomposition,
    rhs::AbstractArray{<:AbstractFloat, 3},
    ::Type{Factor};
    tullio_threading
) -> Any

Solve the linear system based on Decomposition type